+++ it/server/src/test/java/com/walmartlabs/concord/it/server/ProcessIT.java
@@ -9,9 +9,9 @@ package com.walmartlabs.concord.it.server;
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
- * 
+ *
  *      http://www.apache.org/licenses/LICENSE-2.0
- * 
+ *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -21,10 +21,23 @@ package com.walmartlabs.concord.it.server;
  */
 
 import com.googlecode.junittoolbox.ParallelRunner;
+import com.walmartlabs.concord.server.api.org.OrganizationEntry;
+import com.walmartlabs.concord.server.api.org.OrganizationResource;
+import com.walmartlabs.concord.server.api.org.ResourceAccessEntry;
+import com.walmartlabs.concord.server.api.org.ResourceAccessLevel;
+import com.walmartlabs.concord.server.api.org.project.ProjectEntry;
+import com.walmartlabs.concord.server.api.org.project.ProjectOperationResponse;
+import com.walmartlabs.concord.server.api.org.project.ProjectVisibility;
+import com.walmartlabs.concord.server.api.org.team.*;
 import com.walmartlabs.concord.server.api.process.*;
 import com.walmartlabs.concord.server.api.project.CreateProjectResponse;
-import com.walmartlabs.concord.server.api.org.project.ProjectEntry;
 import com.walmartlabs.concord.server.api.project.ProjectResource;
+import com.walmartlabs.concord.server.api.security.apikey.ApiKeyResource;
+import com.walmartlabs.concord.server.api.security.apikey.CreateApiKeyRequest;
+import com.walmartlabs.concord.server.api.security.apikey.CreateApiKeyResponse;
+import com.walmartlabs.concord.server.api.user.CreateUserRequest;
+import com.walmartlabs.concord.server.api.user.UserResource;
+import com.walmartlabs.concord.server.api.user.UserType;
 import org.junit.Ignore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
@@ -395,5 +408,98 @@ public class ProcessIT extends AbstractServerIT {
             }
         }
         assertNotNull(p);
+    }
+
+    @Test(timeout = 60000)
+    public void testGetProcessForChildIds() throws Exception {
+
+        // ---
+        byte[] payload = archive(ProcessIT.class.getResource("processWithChildren").toURI());
+
+        ProcessResource processResource = proxy(ProcessResource.class);
+        StartProcessResponse parentSpr = processResource.start(new ByteArrayInputStream(payload), null, false, null);
+
+        // ---
+
+        waitForCompletion(processResource, parentSpr.getInstanceId());
+
+        ProcessEntry processEntry = processResource.get(parentSpr.getInstanceId());
+        assertEquals(processEntry.getChildrenIds().size(), 3);
+    }
+
+    @Test(timeout = 60000)
+    public void testGetAllProcessesForChildIds() throws Exception {
+        // create a new org
+
+        String orgName = "org_" + randomString();
+
+        OrganizationResource orgResource = proxy(OrganizationResource.class);
+        orgResource.createOrUpdate(new OrganizationEntry(orgName));
+
+        // add the user A
+
+        UserResource userResource = proxy(UserResource.class);
+
+        String userAName = "userA_" + randomString();
+        userResource.createOrUpdate(new CreateUserRequest(userAName, UserType.LOCAL));
+
+        ApiKeyResource apiKeyResource = proxy(ApiKeyResource.class);
+        CreateApiKeyResponse apiKeyA = apiKeyResource.create(new CreateApiKeyRequest(userAName));
+
+        // create the user A's team
+
+        String teamName = "team_" + randomString();
+
+        TeamResource teamResource = proxy(TeamResource.class);
+        CreateTeamResponse ctr = teamResource.createOrUpdate(orgName, new TeamEntry(teamName));
+
+        teamResource.addUsers(orgName, teamName, Collections.singleton(new TeamUserEntry(userAName, TeamRole.MEMBER)));
+
+        // switch to the user A and create a new private project
+
+        setApiKey(apiKeyA.getKey());
+
+        String projectName = "project_" + randomString();
+
+        com.walmartlabs.concord.server.api.org.project.ProjectResource projectResource = proxy(com.walmartlabs.concord.server.api.org.project.ProjectResource.class);
+        ProjectOperationResponse por = projectResource.createOrUpdate(orgName, new ProjectEntry(projectName, ProjectVisibility.PRIVATE));
+
+        // grant the team access to the project
+
+        projectResource.updateAccessLevel(orgName, projectName, new ResourceAccessEntry(ctr.getId(), orgName, teamName, ResourceAccessLevel.READER));
+
+        //Start a process with zero child
+
+        byte[] payload = archive(ProcessIT.class.getResource("process").toURI());
+        Map<String, Object> input = new HashMap<>();
+        input.put("archive", payload);
+        input.put("org", orgName);
+        input.put("project", projectName);
+
+        StartProcessResponse singleNodeProcess = start(input);
+        ProcessResource processResource = proxy(ProcessResource.class);
+        waitForCompletion(processResource, singleNodeProcess.getInstanceId());
+
+        // Start a process with children
+
+        payload = archive(ProcessIT.class.getResource("processWithChildren").toURI());
+        input = new HashMap<>();
+        input.put("archive", payload);
+        input.put("org", orgName);
+        input.put("project", projectName);
+
+        StartProcessResponse parentSpr = start(input);
+        waitForCompletion(processResource, parentSpr.getInstanceId());
+
+        // ---
+
+        List<ProcessEntry> processEntry = processResource.list(por.getId(), null, null, 10);
+        for (ProcessEntry pe : processEntry) {
+            if (pe.getInstanceId().equals(singleNodeProcess.getInstanceId())) {
+                assertTrue(pe.getChildrenIds().isEmpty());
+            } else if (pe.getInstanceId().equals(parentSpr.getInstanceId())) {
+                assertEquals(3, pe.getChildrenIds().size());
+            }
+        }
     }
 }